requirements.txt
google.generativeai
pypdf
streamlit
python-dotenv